+++ src/views/paissa/client.ts
@@ -1,13 +1,5 @@
 import {PAISSADB_BASE, PAISSADB_WS_URL} from "@/utils/constants";
-import {
-    DistrictDetail,
-    OpenPlotDetail,
-    PlotUpdate,
-    SoldPlotDetail,
-    WorldDetail,
-    WorldSummary,
-    WSMessage
-} from "@/utils/paissa/types";
+import {OpenPlotDetail, PlotUpdate, SoldPlotDetail, WorldDetail, WorldSummary, WSMessage} from "@/views/paissa/types";
 import axios from "axios";
 
 interface PlotLoc {
